FIXED: Bytes ratio percentage always 0% FIXED: No such task error when clearing from details.

+++ src/org/jared/synodroid/common/protocol/v22/DSHandlerDSM22.java
@@ -356,7 +356,7 @@ class DSHandlerDSM22 implements DSHandler {
 					}
 					result.bytesRatio = 0;
 					if (result.bytesDownloaded != 0) {
-						result.bytesRatio = (int) ((result.bytesUploaded / result.bytesDownloaded) * 100);
+						result.bytesRatio = (int)(((float)result.bytesUploaded / (float)result.bytesDownloaded) * 100);
 					}
 				}
 				if (data.has("seedelapsed"))
